wb_parser_bot/requirements.txt
python-telegram-bot==21.6
requests==2.31.0
openpyxl==3.1.2
urllib3==2.2.1
